requirements.txt
pydantic>=2.7,<3.0
pyautogen>=0.2.35
openai>=1.66.2,<1.100
gradio>=5.0,<6.0
python-dotenv>=1.0,<2.0
